feat: created DashboardPage and new HomePage
